requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
chromadb
sentence-transformers
wikipedia
python-dotenv
tiktoken
